refactor: use new style in container
