+++ src/ts/status/signin/selectors/authenticated.test.tsx
@@ -1,5 +1,5 @@
 import { createUserId } from "@/domains/user";
-import { RecoilRoot } from "recoil";
+import { MutableSnapshot, RecoilRoot } from "recoil";
 import currentUserState from "../atoms/current-user";
 import authenticatedSelector from "./authenticated";
 import { render } from "@testing-library/react";
@@ -25,7 +25,7 @@ test("user is not authenticated if state is default", async () => {
 });
 
 test("user is authenticated if user is setted", async () => {
-  const snapshot = ({ set }) => {
+  const snapshot = ({ set }: MutableSnapshot) => {
     set(currentUserState, (prev) => ({ ...prev, id: createUserId("1"), name: "name" }));
   };
